+++ slang/lib/builder/builder/build_model_config_builder.dart
@@ -11,6 +11,7 @@ extension BuildModelConfigBuilder on RawConfig {
       stringInterpolation: stringInterpolation,
       maps: maps,
       pluralAuto: pluralAuto,
+      pluralParameter: pluralParameter,
       pluralCardinal: pluralCardinal,
       pluralOrdinal: pluralOrdinal,
       contexts: contexts,